[Logger] Added automatic post-write delay
- calling several times logging methods may cause unexpected behaviour
to occur (since write is done asynchronously)
- The delay in ms can be specified with a new constructor taking the
delay value as parameter (When calling the constructor with no delay)
parameter, the delay is set to "no delay"

+++ Logger/src/core/Logger.h
@@ -18,6 +18,8 @@
 #define LOGGER_h
 
 #include <Arduino.h>
+#include <inttypes.h>
+
 /**
  * This (abstract) class allows to log messages
  */
@@ -43,7 +45,7 @@ public:
    * @return logging success status
    */
   virtual boolean
-  logMessage(byte messageBytes[], int amount) = 0;
+  logMessage(byte messageBytes[], uint16_t amount) = 0;
 
   /**
    * Clear log content (optional).
